submission fixed, metrics fixed
