+++ metrics.py
@@ -1,39 +1,22 @@
 from tqdm import tqdm
 
 import numpy as np
-from scipy import ndimage
 from sklearn.metrics.pairwise import pairwise_distances
 
-
-def decompose(mask):
-    labeled, nr_true = ndimage.label(mask)
-    masks = []
-    for i in range(1, nr_true + 1):
-        msk = labeled.copy()
-        msk[msk != i] = 0.
-        msk[msk == i] = 255.
-        masks.append(msk)
-    if not masks:
-        return [mask]
-    else:
-        return masks
+from utils import decompose
 
 
-def iou(im1, im2, empty_score=1.0):
-    im1 = np.asarray(im1).astype(np.bool)
-    im2 = np.asarray(im2).astype(np.bool)
-
-    if im1.shape != im2.shape:
-        raise ValueError("Shape mismatch: im1 and im2 must have the same shape.")
-
-    im_sum = im1.sum() + im2.sum()
-    if im_sum == 0:
-        return empty_score
-
-    # Compute Dice coefficient
-    intersection = np.logical_and(im1, im2)
-
-    return intersection.sum() / im_sum
+def iou(gt, pred):
+    gt[gt > 0] = 1.
+    pred[pred > 0] = 1.
+    intersection = gt * pred
+    union = gt + pred
+    union[union > 0] = 1.
+    intersection = np.sum(intersection)
+    union = np.sum(union)
+    if union == 0:
+        union = 1e-09
+    return intersection / union
 
 
 def compute_ious(gt, predictions):
@@ -64,7 +47,10 @@ def compute_eval_metric(gt, predictions):
 def intersection_over_union(y_true, y_pred):
     ious = []
     for y_t, y_p in tqdm(list(zip(y_true, y_pred))):
-        ious.append(np.mean(compute_ious(y_t, y_p)))
+        iou = compute_ious(y_t, y_p)
+        iou_mean = 1.0 * np.sum(iou) / iou.shape[0]
+        ious.append(iou_mean)
+
     return np.mean(ious)
